% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairedPulseTMS.R
\name{NORMALISE_PP_to_SP_forDataframe}
\alias{NORMALISE_PP_to_SP_forDataframe}
\title{NORMALISE_PP_to_SP}
\usage{
NORMALISE_PP_to_SP_forDataframe(data, format, number_of_triggers, normalise)
}
\arguments{
\item{data}{Single participant data, including their ID}

\item{format}{How are the data shaped? "Long" or "Wide"}

\item{number_of_triggers}{How many total measures do you have? For example, 4 means two single-pulse and two paired-pulse}

\item{normalise}{What type of normalisation do you want for the paired-pulse data? Can choose from five options: (1) (PP_tX / SP_tX); (2) (PP_tX - SP_tX) / SP_tX; (3) (PP_tX/SP_tX) / (PP_t0 / SP_t0); (4) natural_log(PP_tX / SP_tX); (5) (PP_tX / SP_t0). 'tX' means that this values changes based on the time-point (it could be t_1, t_2, T_3 etc.). 't0' means that this value is fixed at the baseline/starting value}
}
\value{
A file in wide format with all the single- and paired-pulse data normalised
}
\description{
This function is a wrapper for 'SRTT_analysis' to handle dataframes as input data
}
\details{
The functions goes over the 'data' file.
We expect 'data' to have one column for each subject, with ID on top.
We handle also the option where the format is wide
The first row has subject's ID ("ID")
All the other rows are raw excitability measures, which alternate between single-pulse and paired-pulse (e.g., ID, SP_t0, PP_t0, SP_t1, PP_t1, SP_t2 etc.)
}
\examples{
# Basic usage
ID <- LETTERS[1:5]
dataframe_wide <- data.frame(ID = ID, matrix(stats::rnorm(5 * 16, mean = 1, sd = 0.1), nrow = 5, ncol = 16) )
NORMALISE_PP_to_SP_forDataframe(data = dataframe_wide, format = "Wide", number_of_triggers = 16 , normalise = 3)

dataframe_long <- as.data.frame(t(dataframe_wide))
NORMALISE_PP_to_SP_forDataframe(data = dataframe_long, format = "Long", number_of_triggers = 16 , normalise = 4)
}
