% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairedPulseTMS.R
\name{plot_data_for_Dataframes}
\alias{plot_data_for_Dataframes}
\title{plot_data_for_Dataframes}
\usage{
plot_data_for_Dataframes(
  data,
  format = "Wide",
  col_line = "grey",
  col_dots = "grey",
  error_measure = "se"
)
}
\arguments{
\item{data}{Dataframe including participant data, and their ID}

\item{format}{How are the data shaped? "Long" or "Wide"}

\item{col_line}{What color do you want to use for the line? (e.g., "black")}

\item{col_dots}{What color do you want to use for the dots? (e.g., "grey")}

\item{error_measure}{How do you want to calculate the group-wise error measure? "sd" for standard deviation or "se" for standard error of the mean}
}
\value{
Three figures returned in a list and modifiable
}
\description{
This function handles dataframes with one or multiple IDs as input data
this is a function to visualize excitability measures
each dot is a response
X-axis = type of excitability measures (TS_alone or CS & TS)
Y-axis = excitability amplitude
}
\examples{
# Basic usage
ID <- LETTERS[1:5]
dataframe_wide <- data.frame(ID = ID, matrix(stats::rnorm(5 * 16, mean = 1, sd = 0.5), nrow = 5, ncol = 16) )
data <- NORMALISE_PP_to_SP_forDataframe(data = dataframe_wide, format = "Wide", number_of_triggers = 16 , normalise = 5)
plots <- plot_data_for_Dataframes(data, format = "Wide", col_line = "grey", col_dots = "grey", error_measure = "se")
plot1 <- plots$plot1
plot1 + ggplot2::geom_hline(yintercept = 1, col = "red")
plots$plot2
plots$plot3
data_wide <- data[1, ]
plots <- plot_data_for_Dataframes(data_wide, format = "Wide", col_line = "grey", col_dots = "grey", error_measure = "se")
plots$plot3
}
